(tail): Explicitly cast to long the N_UNITS argumennt in call to tail_lines.  Otherwise, NetBSD lost because N_UNITS (of type off_t, a 64-bit type on NetBSD) wasn't converted to the target type, long.  Reported by Dan Hagerty (hag@gnu.ai.it.edu).

+++ src/tail.c
@@ -447,7 +447,7 @@ tail (filename, fd, n_units)
      off_t n_units;
 {
   if (count_lines)
-    return tail_lines (filename, fd, n_units);
+    return tail_lines (filename, fd, (long) n_units);
   else
     return tail_bytes (filename, fd, n_units);
 }
